rename TableOfContentsPlugin module
