+++ packages/lexical-react/flow/LexicalTableOfContents.js.flow
@@ -7,15 +7,7 @@
  * @flow strict
  */
 
-import type {HeadingTagType} from '@lexical/rich-text';
-import type {LexicalEditor, NodeKey} from 'lexical';
+import {TableOfContentsPlugin} from '@lexical/react/LexicalTableOfContentsPlugin';
 
-declare export function LexicalTableOfContents({
-  children: (
-    tableOfContents: Array<[NodeKey, string, HeadingTagType]>,
-    editor: LexicalEditor,
-  ) => React$Node,
-}): React$Node;
-
-/** @deprecated use the named export {@link LexicalTableOfContents} */
-declare export default typeof LexicalTableOfContents;
+/** @deprecated use the named export {@link LexicalTableOfContentsPlugin} */
+declare export default typeof TableOfContentsPlugin;
