+++ packages/lexical-react/flow/LexicalTableOfContentsPlugin.js.flow
@@ -1,0 +1,18 @@
+/**
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow strict
+ */
+
+import type {HeadingTagType} from '@lexical/rich-text';
+import type {LexicalEditor, NodeKey} from 'lexical';
+
+declare export function TableOfContentsPlugin({
+  children: (
+    tableOfContents: Array<[NodeKey, string, HeadingTagType]>,
+    editor: LexicalEditor,
+  ) => React$Node,
+}): React$Node;
